chore: rename hash calculator to follow naming schema
