+++ packages/hash-stream-node/src/HashCalculator.ts
